# Reduce build targets to save building time