--- a/CHT/ModuleCHT.C
+++ b/CHT/ModuleCHT.C
@@ -1,0 +1,7 @@
+// Central inlude file to reduce the number of build targets and build time
+#include "Temperature.C"
+#include "KappaEffective.C"
+#include "HeatFlux.C"
+#include "HeatTransferCoefficient.C"
+#include "SinkTemperature.C"
+#include "CHT.C"