--- a/FF/ModuleFF.C
+++ b/FF/ModuleFF.C
@@ -1,0 +1,6 @@
+// Central inlude file to reduce the number of build targets and build time
+#include "FF.C"
+#include "Velocity.C"
+#include "VelocityGradient.C"
+#include "Pressure.C"
+#include "PressureGradient.C"