--- a/FSI/ModuleFSI.C
+++ b/FSI/ModuleFSI.C
@@ -1,0 +1,7 @@
+// Central inlude file to reduce the number of build targets and build time
+#include "FSI.C"
+#include "ForceBase.C"
+#include "Force.C"
+#include "Stress.C"
+#include "Displacement.C"
+#include "DisplacementDelta.C"