--- a/changelog-entries/301.md
+++ b/changelog-entries/301.md
@@ -1,0 +1,1 @@
+- Reduced building time by grouping together compilation units of each module [#301](https://github.com/precice/openfoam-adapter/pull/301)